requirements.txt
accelerate==1.9.0
annotated-types==0.7.0
anyio==4.9.0
blis==1.3.0
boto3==1.39.17
botocore==1.39.17
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
comtypes==1.4.10
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
duckduckgo_search==8.0.1
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fonttools==4.57.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflect==7.5.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.1.1
openai==1.98.0
opencv-contrib-python-headless==4.12.0.88
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pillow_heif==0.22.0
preshed==3.0.10
primp==0.15.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pypiwin32==223
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzbar==0.1.9
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
roboflow==1.1.61
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.13.1
thinc==8.3.6
tifffile==2025.6.11
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.52.3
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.114
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
